- Add new Expectation engine. This thing allows applying selectors on the dom that should be applied to widgets in runtime. Basically, it's like this:
widget.watch('select:focused option:selected', function(widget, state) { 
  // state will be true when widget matches the selector
  // or false when it no longer matches it.
  widget[state ? 'addEvents' : 'removeEvents'](...)
});

CSS matcher was revamped to use this technique. Slick is now doing all the dirty work so we have all the cool selectors Slick supports right in LSD-CSS! New combinators and pseudo classes... Just tasty.

this.use(['#a', '#b', 'menu[type=menu]', function(a, b, menu, state) {
  // use maps array of selector to arguments. 
});

Actions are redone to work in asycnronous way. So you can drop-in handles for the draggable widget and they will be picked up on the fly and the events will be asigned. 

I'll write about it more later.

Few more bugs to be ironed out for release.

- Multiple stylesheets support. They are now classes, not a singleton. Sheet.js is smaller now and 100% more clean. 

- Porting things to 1.3... Slowly.

- Accepted a pull request from rwz & paxa @ github. Thank you very much guys, I love you for this. Particular pull makes mootools-ext fully 1.3 compatible and also adds docs (in haml! can you believe it?).

  Looking forward to further collaboration. 

- Added glyph shadow layer to all buttons. Cuute.

- Behaviours are now gone (as separate entity), use expectations instead.