Wire in commands this time for real. Even if it's not so useful as the example i demoed in that commit message 2 weeks ago, but atleast it is functional. Checkboxes, radiobuttons and buttons rely soloely on commands.
When radiobutton is hit, instead of looping through dom nodes to find the right other radiobutton to uncheck, we take a shortcut and look right through radiogroups that are defined on document by Command.Radio module.

All those widgets have a method "click" that is similar to those in regular DOM's. They take actions as if button (checkbox, radiobutton) were hit for real and change the state if needed. The difference between LSD command .click and dom link.click is that LSD fires up click event on widget.